Putting back arrow to the startexperiment page. Basically moved one function and one line of code from ImageActivity to MainActivity.
